+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.0](https://github.com/seamapi/python/compare/v1.2.0...v1.3.0) (2022-09-19)
+
+
+### Features
+
+* Merge pull request [#42](https://github.com/seamapi/python/issues/42) from seamapi/implement-delete-connected-account ([c4b55b5](https://github.com/seamapi/python/commit/c4b55b5722774f6dc700679ecd03fbc12028eda0)), closes [#27](https://github.com/seamapi/python/issues/27)
+
+## [1.2.0](https://github.com/seamapi/python/compare/v1.1.5...v1.2.0) (2022-09-16)
+
+
+### Features
+
+* Merge pull request [#41](https://github.com/seamapi/python/issues/41) from seamapi/implement-device-update ([bd94e17](https://github.com/seamapi/python/commit/bd94e17bab7657d31e2f58049deb136b5d2a58a2)), closes [#29](https://github.com/seamapi/python/issues/29)
+
 ### [1.1.5](https://github.com/seamapi/python/compare/v1.1.4...v1.1.5) (2022-09-13)